Restructure knative reconciler pkg
