+++ pkg/credentials/gcs/gcs_secret.go
