+++ pkg/reconciler/knative/knative_suite_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package ksvc
+package knative
 
 import (
 	pkgtest "github.com/kubeflow/kfserving/pkg/testing"
